improved build and packaging process

+++ package/build.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [ "$#" != "1" ]; then
+    echo "Illegal number of parameters, write only a number in the following format _._._"
+    exit 0
+fi
+
 cd ..
 ./paket restore
 rm -r package/Release
@@ -7,7 +12,7 @@ rm -r src/bin/Release
 msbuild src/ /property:Configuration=Release 
 cd package
 cp -r ../src/bin/Release .
-cp ../packages/FSharp.Core/lib/net45/FSharp.Core.dll Release/
-cp fstar.exe.config Release/
 
+touch ./Release/.npmignore
+npm version $1
 npm pack .
